Support CInputStream_NetCDF input_file csv list
Previously, was processing a ton of individual files
each with a call to fpf. Although admittedly fpf was
not the slow part, this was the best way to realize
an accelleration.

Can avoid many calls through the Python wrapper,
instead, keep the calls within fpf, have all
the files write to one writer node, and then
only one call to ccsdspy after.